requirements_pinf.txt
Babel==1.3
Django==1.8.2
JPype1==0.6.0
JayDeBeApi==0.2.0
Jinja2==2.7.3
Markdown==2.6
MarkupSafe==0.23
MySQL-python==1.2.5
PyYAML==3.11
Pygments==2.0.2
SQLAlchemy==1.0.4
Sphinx==1.3.1
WebOb==1.4.1
alabaster==0.7.4
argparse==1.2.1
async==0.6.1
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
blist==1.3.6
cassandra-driver==2.5.1
certifi==2015.04.28
cqlengine==0.20.0
crispy-forms-foundation==0.3.5
django-ajax
django-cassandra-engine==0.3.3
django-countries==2.1.2
django-crispy-forms==1.4.0
django-dajaxice==0.7
django-debug-toolbar==1.2.2
django-extensions==1.4.0
django-filter==0.8
django-form-utils==1.0.2
django-geojson==2.8.1
django-inplaceedit==1.3.0
django-jsonfield==0.9.13
django-jsonview==0.4.3
django-mongodb-engine==0.5.2
django-registration==1.0
django-split-json-widget==0.2
django-tables2==1.0.4
django-tables2-reports==0.0.9
django-toolbox==0.1
django-utils==0.0.2
djangoappengine==1.6.3
djangorestframework==3.0.5
djangotoolbox==1.6.2
djorm-ext-pgfulltext==0.9.2
docker-py==0.5.3
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
fig==1.0.1
freeseer==3.0.1
functools32==3.2.3-1
futures==3.0.3
happybase==0.8
ipython==3.1.0
jdcal==1.0
jsonfield==1.0.3
jsonschema==2.5.1
kazoo==2.0
meld3==1.0.0
mistune==0.5.1
numpy==1.9.1
openpyxl==2.1.2
pandas==0.15.2
psycopg2==2.5.4
ptyprocess==0.5
pyExcelerator==0.6.4.1
pyaccumulo==1.5.0.8
pymongo==2.7.2
python-dateutil==2.4.0
pytz==2014.10
pyzmq==14.6.0
querystring-parser==1.2.0
requests==2.7.0
singledispatch==3.4.0.3
six==1.9.0
snowballstemmer==1.2.0
sphinx-rtd-theme==0.1.8
sqlparse==0.1.14
supervisor==3.1.3
tablib==0.10.0
terminado==0.5
texttable==0.8.3
thrift==0.9.1
tornado==4.2
webapp2==2.5.2
websocket-client==0.11.0
wsgiref==0.1.2
xlrd==0.9.3
xlwt==0.7.5
